+++ packages/stylelint/src/miniapp.ts
@@ -5,8 +5,12 @@ export default {
   ...MINIAPP_SHARED_CONFIG,
   overrides: [
     {
-      files: ['*.wxss'],
+      files: ['**/*.wxss'],
       customSyntax: 'postcss',
+    },
+    {
+      files: ['**/*.wxml'],
+      customSyntax: 'postcss-html',
     },
   ],
 } satisfies Config