+++ packages/stylelint/src/vue-scss.ts
@@ -1,0 +1,6 @@
+import type { Config } from 'stylelint'
+
+export default {
+  // https://github.com/ota-meshi/stylelint-config-standard-vue#with-scss
+  extends: ['stylelint-config-standard-vue/scss'],
+} satisfies Config