+++ packages/stylelint/src/vue.ts
@@ -1,12 +1,6 @@
 import type { Config } from 'stylelint'
 
 export default {
-  // https://github.com/ota-meshi/stylelint-config-recommended-vue
-  extends: ['stylelint-config-recommended-vue'],
-  overrides: [
-    {
-      files: ['**/*.vue'],
-      customSyntax: 'postcss-html',
-    },
-  ],
+  // https://github.com/ota-meshi/stylelint-config-standard-vue
+  extends: ['stylelint-config-standard-vue'],
 } satisfies Config